% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_viz.R
\name{simple_viz}
\alias{simple_viz}
\title{A simple visualisation for MACS results}
\usage{
simple_viz(bed_file, track_file)
}
\arguments{
\item{bed_file}{BED file for peaks detected by MACS}

\item{track_file}{BDG file created by MACS}
}
\value{
A ggplot object
}
\description{
A simple visualisation for MACS results
}
